requirements.txt
discord.py
pylint
pytest
coverage
python-dotenv

